Create README.md
1 - what the homework was about in your own words
The homework involved developing a general idea on a specific and important issue using our knowledge, literature, and theory. I then needed to formulate alternative and null hypotheses based on this idea. Finally, I assessed the strengths and weaknesses of the experimental design we plan to use for testing these hypotheses.

2 - who you worked with
John Uponi and I worked together for this assignment.
3 - what things were you responsible
We divided part 2 (the hypothesis section) among us, with each person working on two questions individually. We then shared our findings and discussed them together.
For the Philly parks project, we collaborated on coding, interpretation, and creating captions for the figures, working together over Zoom. We held a Zoom meeting for over two hours to share our ideas and responses to the questions.